feature: Increase precision of the unit input step constraint
I originally set it to 2 decimal places because I felt like that was a sufficient level of precision for this piece of data. It probably is, but the constraint was still causing problems. When you enter ounces the application converts it to pounds and it may end up being a number with greater than 2 decimal places. This means, if you're using the enter again feature, the automatically filled in data may be invalid because the units number may be too precise.

I increased the step to a very large level of precision that shouldn't cause problems going forward.