feature: Improve name of task
Changing the name of this task from `review_app:prime` to
`db:add_dummy_data` makes sense because I don't only use this task
for populating the review apps. That was the original purpose, but now
I use it to populate my local development database as well. I'm updating
the name to reflect that.