feature: Move expiration dates below storage types
Having the expiration dates above the storage types works fine as long
as you don't have very many expiring items. However, if you do have a
lot of expiring items, it's a pain to scroll through them all to get to your
storage types.

I think it also just makes more sense to put the expiration dates below
the storage types because the storage types are more important and
are more likely to be what the user is interested in.

I originally liked the idea of putting the expiration dates at the top
because then they're the first thing you see and it reminds you what
food items you should focus on rotating, but it just doesn't work as
well to have them at the top.